Fix bus power problems in onewire
See #676

The 1-wire pin was being set back to output low, which reset all 1-wire
devices on the bus.

Change the power == 0 behaviour to tri-state the pin, as I believe was
intended. Make this not the default, because the bus should be high
when a bit/byte is not being read/written. A low bus for a long time
is a reset condition.

Signed-off-by: Nick Andrew <nick@nick-andrew.net>